+++ src/utils.js
@@ -4,13 +4,6 @@
 
 var slice = Array.prototype.slice;
 
-function freeze(obj_or_arr, unless) {
-  if (unless !== true && Object.freeze) {
-    Object.freeze(obj_or_arr);
-  }
-  return obj_or_arr;
-}
-
 function mixin(target, source, overwrite) {
   for (var k in source) {
     if (source.hasOwnProperty(k)) {